이미지 추가

+++ DirectX2DProejct/Engine_SOURCE/CConstantBuffer.cpp
@@ -15,7 +15,7 @@ namespace sh::graphics
 
     bool CConstantBuffer::Create(size_t size)
     {
-        desc.ByteWidth = size;
+        desc.ByteWidth = (UINT)size;
         desc.BindFlags = D3D11_BIND_FLAG::D3D11_BIND_CONSTANT_BUFFER;
         desc.Usage = D3D11_USAGE::D3D11_USAGE_DYNAMIC;
         desc.CPUAccessFlags = D3D11_CPU_ACCESS_WRITE;
